+++ DelvUI/changelog.md
@@ -8,6 +8,7 @@ Features:
 - Added "Start Angle" setting for the circular GCD Indicator.
 - Added sub-options to "Hide only JobPack HUD outside of combat" to always show in duties and/or when weapons are drawn.
 - Implement new tag `[time-till-max-gp]`, this will show you the time until your GP hits max again.
+- Added Raise Tracker and Invuln Tracker to the party frames preview mode.
 
 Fixes: 
 - Fixed Bard's Soul Voice threshold not working.